Fix Cookie Decryption Errors
This is an amendment to the previous fix for the notice: "Undefined offset: 1".  $_COOKIE returns the cookie with spaces in the value in place of + signs, causing the cookie decryption to fail.  Using HTTP_COOKIE returns the correct cookie value and the strpos() check blocks bad cookie values, stopping the notice.