Fix PerimeterxOriginalTokanValidator file name
I was seeing issues where attempting to use this class where throwing PHP Fatal errors

PHP Fatal error:  Class 'Perimeterx\PerimeterxOriginalTokenValidator' not found in
path/to/vendor/perimeterx/php-sdk/src/PerimeterxCookieValidator.php on line 53

It seems the autoloader does not like class and filename capitalization to be different.
I'm not sure if this is an issue with outdated PHP (5.6), outdated composer, etc, so this
may not be a widespread issue.